snapshot of project "reflex", label t20131209

+++ initscan.c
@@ -3100,7 +3100,7 @@ input(void)
 	    *yy_c_buf_p = '\0';
 
 	else {			/* need more input */
-	    int offset = yy_c_buf_p - yytext_ptr;
+	    int offset = (int) (yy_c_buf_p - yytext_ptr);
 	    ++yy_c_buf_p;
 
 	    switch (yy_get_next_buffer()) {
